removed the example files
