--- src/BaseBoard/BaseBoard.scad
@@ -1,1 +1,0 @@
-difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){translate([0, 0, 0])cube([1100, 100,15]);translate([100, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([100, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([100, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([100, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([100, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([100, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([100, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([100, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([200, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([200, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([200, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([200, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([200, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([200, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([200, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([200, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([300, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([300, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([300, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([300, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([300, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([300, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([300, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([300, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([400, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([400, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([400, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([400, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([400, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([400, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([400, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([400, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([500, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([500, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([500, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([500, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([500, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([500, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([500, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([500, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([600, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([600, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([600, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([600, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([600, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([600, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([600, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([600, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([700, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([700, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([700, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([700, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([700, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([700, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([700, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([700, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([800, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([800, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([800, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([800, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([800, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([800, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([800, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([800, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([900, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([900, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([900, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([900, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([900, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([900, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([900, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([900, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([1000, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([1000, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([1000, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([1000, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([1000, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([1000, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([1000, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([1000, 90, 15])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}}
