--- src/Blob/Blob.json
@@ -1,54 +1,0 @@
-[
-    {
-        "name": "cube",
-        "type": "add",
-        "side": null,
-        "x": 0,
-        "y": 0,
-        "z": 0,
-        "x_size": 10,
-        "y_size": 10,
-        "z_size": 10,
-        "center": false
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 4,
-        "y": 4,
-        "z": 0.0,
-        "side": "BOTTOM",
-        "diameter": 2,
-        "depth": 12
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 4,
-        "y": 6,
-        "z": 0.0,
-        "side": "BOTTOM",
-        "diameter": 2,
-        "depth": 12
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 6,
-        "y": 4,
-        "z": 0.0,
-        "side": "BOTTOM",
-        "diameter": 2,
-        "depth": 12
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 6,
-        "y": 6,
-        "z": 0.0,
-        "side": "BOTTOM",
-        "diameter": 2,
-        "depth": 12
-    }
-]
