--- src/Thing/Thing.json
@@ -1,198 +1,0 @@
-[
-    {
-        "name": "cube",
-        "type": "add",
-        "side": null,
-        "x": 0,
-        "y": 0,
-        "z": 0,
-        "x_size": 10,
-        "y_size": 10,
-        "z_size": 26,
-        "center": false
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 4,
-        "y": 4,
-        "z": 26.0,
-        "side": "TOP",
-        "diameter": 2,
-        "depth": 12
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 4,
-        "y": 4,
-        "z": 0.0,
-        "side": "BOTTOM",
-        "diameter": 3,
-        "depth": 12
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 4,
-        "y": 4,
-        "z": 0.0,
-        "side": "BOTTOM",
-        "diameter": 3.1,
-        "depth": 12
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 4,
-        "y": 4,
-        "z": 26.0,
-        "side": "TOP",
-        "diameter": 3.1,
-        "depth": 12
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 4,
-        "y": 6,
-        "z": 26.0,
-        "side": "TOP",
-        "diameter": 2,
-        "depth": 12
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 4,
-        "y": 6,
-        "z": 0.0,
-        "side": "BOTTOM",
-        "diameter": 3,
-        "depth": 12
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 4,
-        "y": 6,
-        "z": 0.0,
-        "side": "BOTTOM",
-        "diameter": 3.1,
-        "depth": 12
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 4,
-        "y": 6,
-        "z": 26.0,
-        "side": "TOP",
-        "diameter": 3.1,
-        "depth": 12
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 6,
-        "y": 4,
-        "z": 26.0,
-        "side": "TOP",
-        "diameter": 2,
-        "depth": 12
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 6,
-        "y": 4,
-        "z": 0.0,
-        "side": "BOTTOM",
-        "diameter": 3,
-        "depth": 12
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 6,
-        "y": 4,
-        "z": 0.0,
-        "side": "BOTTOM",
-        "diameter": 3.1,
-        "depth": 12
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 6,
-        "y": 4,
-        "z": 26.0,
-        "side": "TOP",
-        "diameter": 3.1,
-        "depth": 12
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 6,
-        "y": 6,
-        "z": 26.0,
-        "side": "TOP",
-        "diameter": 2,
-        "depth": 12
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 6,
-        "y": 6,
-        "z": 0.0,
-        "side": "BOTTOM",
-        "diameter": 3,
-        "depth": 12
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 6,
-        "y": 6,
-        "z": 0.0,
-        "side": "BOTTOM",
-        "diameter": 3.1,
-        "depth": 12
-    },
-    {
-        "name": "hole",
-        "type": "cut",
-        "x": 6,
-        "y": 6,
-        "z": 26.0,
-        "side": "TOP",
-        "diameter": 3.1,
-        "depth": 12
-    },
-    {
-        "name": "cube",
-        "type": "cut",
-        "x": 3.3333333333333335,
-        "y": -1,
-        "z": 5,
-        "side": "FRONT",
-        "x_size": 3.3333333333333335,
-        "y_size": 5.333333333333334,
-        "z_size": 16,
-        "center": false
-    },
-    {
-        "name": "cube",
-        "type": "cut",
-        "x": -1,
-        "y": 3.3333333333333335,
-        "z": 5,
-        "side": "LEFT",
-        "x_size": 5.333333333333334,
-        "y_size": 3.3333333333333335,
-        "z_size": 16,
-        "center": false
-    }
-]
