--- src/Thing/Thing.scad
@@ -1,1 +1,0 @@
-difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){difference(){translate([0, 0, 0])cube([10, 10,26]);translate([4, 4, 26.0])rotate([0, 180, 0])cylinder(r=1.0, h=12, $fn=64);translate([4, 4, 0.0])rotate([0, 0, 0])cylinder(r=1.5, h=12, $fn=64);translate([4, 4, 0.0])rotate([0, 0, 0])cylinder(r=1.55, h=12, $fn=64);translate([4, 4, 26.0])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([4, 6, 26.0])rotate([0, 180, 0])cylinder(r=1.0, h=12, $fn=64);translate([4, 6, 0.0])rotate([0, 0, 0])cylinder(r=1.5, h=12, $fn=64);translate([4, 6, 0.0])rotate([0, 0, 0])cylinder(r=1.55, h=12, $fn=64);translate([4, 6, 26.0])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([6, 4, 26.0])rotate([0, 180, 0])cylinder(r=1.0, h=12, $fn=64);translate([6, 4, 0.0])rotate([0, 0, 0])cylinder(r=1.5, h=12, $fn=64);translate([6, 4, 0.0])rotate([0, 0, 0])cylinder(r=1.55, h=12, $fn=64);translate([6, 4, 26.0])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([6, 6, 26.0])rotate([0, 180, 0])cylinder(r=1.0, h=12, $fn=64);translate([6, 6, 0.0])rotate([0, 0, 0])cylinder(r=1.5, h=12, $fn=64);translate([6, 6, 0.0])rotate([0, 0, 0])cylinder(r=1.55, h=12, $fn=64);translate([6, 6, 26.0])rotate([0, 180, 0])cylinder(r=1.55, h=12, $fn=64);translate([3.3333333333333335, -1, 5])cube([3.3333333333333335, 5.333333333333334,16]);translate([-1, 3.3333333333333335, 5])cube([5.333333333333334, 3.3333333333333335,16]);}}}}}}}}}}}}}}}}}}
