--- src/Thlob/Thlob.json
@@ -1,401 +1,0 @@
-[
-    {
-        "part_no": "Thing",
-        "rotmax": [
-            10,
-            10,
-            26
-        ],
-        "moves": [
-            100,
-            90,
-            15
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Blob",
-        "rotmax": [
-            10,
-            10,
-            10
-        ],
-        "moves": [
-            100,
-            90,
-            41
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Thing",
-        "rotmax": [
-            10,
-            10,
-            26
-        ],
-        "moves": [
-            200,
-            90,
-            15
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Blob",
-        "rotmax": [
-            10,
-            10,
-            10
-        ],
-        "moves": [
-            200,
-            90,
-            41
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Thing",
-        "rotmax": [
-            10,
-            10,
-            26
-        ],
-        "moves": [
-            300,
-            90,
-            15
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Blob",
-        "rotmax": [
-            10,
-            10,
-            10
-        ],
-        "moves": [
-            300,
-            90,
-            41
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Thing",
-        "rotmax": [
-            10,
-            10,
-            26
-        ],
-        "moves": [
-            400,
-            90,
-            15
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Blob",
-        "rotmax": [
-            10,
-            10,
-            10
-        ],
-        "moves": [
-            400,
-            90,
-            41
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Thing",
-        "rotmax": [
-            10,
-            10,
-            26
-        ],
-        "moves": [
-            500,
-            90,
-            15
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Blob",
-        "rotmax": [
-            10,
-            10,
-            10
-        ],
-        "moves": [
-            500,
-            90,
-            41
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Thing",
-        "rotmax": [
-            10,
-            10,
-            26
-        ],
-        "moves": [
-            600,
-            90,
-            15
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Blob",
-        "rotmax": [
-            10,
-            10,
-            10
-        ],
-        "moves": [
-            600,
-            90,
-            41
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Thing",
-        "rotmax": [
-            10,
-            10,
-            26
-        ],
-        "moves": [
-            700,
-            90,
-            15
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Blob",
-        "rotmax": [
-            10,
-            10,
-            10
-        ],
-        "moves": [
-            700,
-            90,
-            41
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Thing",
-        "rotmax": [
-            10,
-            10,
-            26
-        ],
-        "moves": [
-            800,
-            90,
-            15
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Blob",
-        "rotmax": [
-            10,
-            10,
-            10
-        ],
-        "moves": [
-            800,
-            90,
-            41
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Thing",
-        "rotmax": [
-            10,
-            10,
-            26
-        ],
-        "moves": [
-            900,
-            90,
-            15
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Blob",
-        "rotmax": [
-            10,
-            10,
-            10
-        ],
-        "moves": [
-            900,
-            90,
-            41
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Thing",
-        "rotmax": [
-            10,
-            10,
-            26
-        ],
-        "moves": [
-            1000,
-            90,
-            15
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "Blob",
-        "rotmax": [
-            10,
-            10,
-            10
-        ],
-        "moves": [
-            1000,
-            90,
-            41
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "red"
-    },
-    {
-        "part_no": "BaseBoard",
-        "rotmax": [
-            1100,
-            100,
-            15
-        ],
-        "moves": [
-            0,
-            0,
-            0
-        ],
-        "rotate": [
-            0,
-            0,
-            0
-        ],
-        "colour": "grey"
-    }
-]
