--- src/Thlob/Thlob.scad
@@ -1,1 +1,0 @@
-translate([100.0, 90.0, 15.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Thing/Thing.stl");translate([100.0, 90.0, 41.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Blob/Blob.stl");translate([200.0, 90.0, 15.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Thing/Thing.stl");translate([200.0, 90.0, 41.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Blob/Blob.stl");translate([300.0, 90.0, 15.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Thing/Thing.stl");translate([300.0, 90.0, 41.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Blob/Blob.stl");translate([400.0, 90.0, 15.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Thing/Thing.stl");translate([400.0, 90.0, 41.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Blob/Blob.stl");translate([500.0, 90.0, 15.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Thing/Thing.stl");translate([500.0, 90.0, 41.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Blob/Blob.stl");translate([600.0, 90.0, 15.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Thing/Thing.stl");translate([600.0, 90.0, 41.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Blob/Blob.stl");translate([700.0, 90.0, 15.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Thing/Thing.stl");translate([700.0, 90.0, 41.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Blob/Blob.stl");translate([800.0, 90.0, 15.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Thing/Thing.stl");translate([800.0, 90.0, 41.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Blob/Blob.stl");translate([900.0, 90.0, 15.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Thing/Thing.stl");translate([900.0, 90.0, 41.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Blob/Blob.stl");translate([1000.0, 90.0, 15.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Thing/Thing.stl");translate([1000.0, 90.0, 41.0])rotate([0.0, 0.0, 0.0])color("red")import("/home/helen/tsolo/cycax/src/Blob/Blob.stl");translate([0.0, 0.0, 0.0])rotate([0.0, 0.0, 0.0])color("grey")import("/home/helen/tsolo/cycax/src/BaseBoard/BaseBoard.stl");
